requirements.txt
python-telegram-bot>=20.7
yt-dlp>=2024.04.09
openai>=1.10.0
